feat: move apps/web to app directory
