+++ apps/web/.gitignore
@@ -32,3 +32,6 @@ yarn-error.log*
 
 # vercel
 .vercel
+
+# contentlayer
+.contentlayer
