+++ apps/web/package.json
@@ -13,8 +13,7 @@
     "@tiptap/pm": "^2.0.3",
     "date-fns": "^2.28.0",
     "gray-matter": "^4.0.3",
-    "next": "^13.0.0",
-    "next-transpile-modules": "^10.0.0",
+    "next": "^13.4.8",
     "outstatic": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
