+++ apps/web/src/components/Analytics.tsx
@@ -1,17 +1,19 @@
+'use client'
 import Script from 'next/script'
 import { GA_TRACKING_ID } from '../lib/analytics'
 
-const Analytics = () => (
-  <>
-    <Script
-      strategy="afterInteractive"
-      src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
-    />
-    <Script
-      id="google-analytics"
-      strategy="afterInteractive"
-      dangerouslySetInnerHTML={{
-        __html: `
+const Analytics = () => {
+  return (
+    <>
+      <Script
+        strategy="afterInteractive"
+        src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
+      />
+      <Script
+        id="google-analytics"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
             window.dataLayer = window.dataLayer || [];
             function gtag(){dataLayer.push(arguments);}
             gtag('js', new Date());
@@ -19,9 +21,10 @@ const Analytics = () => (
               page_path: window.location.pathname,
             });
           `
-      }}
-    />
-  </>
-)
+        }}
+      />
+    </>
+  )
+}
 
 export default Analytics
