--- apps/web/src/pages/_document.tsx
@@ -1,35 +1,0 @@
-import Document, {
-  DocumentContext,
-  DocumentInitialProps,
-  Html,
-  Head,
-  Main,
-  NextScript
-} from 'next/document'
-import HeadTags from '../components/HeadTags'
-
-class MyDocument extends Document {
-  static async getInitialProps(
-    ctx: DocumentContext
-  ): Promise<DocumentInitialProps> {
-    const initialProps = await Document.getInitialProps(ctx)
-
-    return initialProps
-  }
-
-  render() {
-    return (
-      <Html lang="en">
-        <Head>
-          <HeadTags />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-}
-
-export default MyDocument
