+++ apps/web/tailwind.config.js
@@ -3,6 +3,7 @@ const plugin = require('tailwindcss/plugin')
 
 module.exports = {
   content: [
+    './src/app/**/*.{js,ts,jsx,tsx}',
     './src/pages/**/*.{js,ts,jsx,tsx}',
     './src/components/**/*.{js,ts,jsx,tsx}'
   ],
